Register, Login, Logout and Admin page management

+++ API/controllers/user.controller.ts
@@ -4,6 +4,19 @@ import { Request, Response } from 'express';
 import user from '../db/user';
 import { User } from '../models/user';
 
+const validateToken = (req: Request, res: Response) => {
+    const { token } = req.body;
+    const secretKey = process.env.JWT_SECRET || 'default_secret_key';
+
+    jwt.verify(token, secretKey, (err) => {
+    if (err) {
+        return res.status(200).send({ message: 'Invalid token' });
+    } else {
+        return res.status(200).send({ message: 'Valid token' });
+    }
+    });
+}
+
 const login = async (req: Request, res: Response) => {
     try {
         const { username, email, password } = req.body;
@@ -12,7 +25,8 @@ const login = async (req: Request, res: Response) => {
             username: '',
             email: '',
             password: '',
-            role: ''
+            role: '',
+            image_url: ''
         };
 
         if(username) {
@@ -41,11 +55,45 @@ const login = async (req: Request, res: Response) => {
         };
 
         const token = jwt.sign(tokenPayload, secretKey, { expiresIn: '2h' });
+        const connectedUser: any = {
+            id: existingUser.id,
+            username: existingUser.username,
+            role: existingUser.role,
+            image_url: existingUser.image_url
+        }
 
         res.status(200).send({
             message: 'Login successful',
-            token: token
+            token: token,
+            connectedUser: connectedUser
         });
+    } catch (error) {
+        console.error('Error occurred:', error);
+        res.status(500).send({ message: 'Internal Server Error' });
+    }
+};
+
+const logout = (req: Request, res: Response) => {
+    try {
+        const { token } = req.body;
+        
+        if (!token) {
+            return res.status(401).send({ message: 'Invalid or missing token' });
+        }
+
+        const secretKey = process.env.JWT_SECRET || 'default_secret_key';
+        const decodedToken = jwt.decode(token);
+
+        if (!decodedToken) {
+            return res.status(403).send({ message: 'Invalid token' });
+        }
+
+        const expiredToken = jwt.sign({ 
+            tokenPayload: decodedToken.tokenPayload, 
+            expiresIn: Math.floor(Date.now() / 1000) - 1
+        }, secretKey);
+
+        res.status(200).send({ message: 'Logout successful' });
     } catch (error) {
         console.error('Error occurred:', error);
         res.status(500).send({ message: 'Internal Server Error' });
@@ -54,9 +102,13 @@ const login = async (req: Request, res: Response) => {
 
 const getAll = (req: Request, res: Response) => {
     user.selectAll().then(users => {
+        let returnedUsers: any[] = [];
+        for (const gotUser of users) {
+            returnedUsers.push({id:gotUser.id, username: gotUser.username, email: gotUser.email, role: gotUser.role, image_url: gotUser.image_url})
+        }
         res.status(200).send({
             message: 'Users List',
-            result: users
+            result: returnedUsers
         })
     }).catch(err => {
         res.status(500).send({
@@ -70,6 +122,10 @@ const getUserById = async (req: Request, res: Response) => {
     try {
         const userId = Number(req.params.id);
 
+        if (!(req.isAuth && req.user.id === userId)) {
+            return res.status(400).send({ message: 'You are not authorized to perform this action.' });
+        } 
+
         if (isNaN(userId)) {
             return res.status(400).send({ message: 'Invalid user ID' });
         }
@@ -77,9 +133,16 @@ const getUserById = async (req: Request, res: Response) => {
         const userResult = await user.getUserById(userId);
 
         if (userResult) {
+            const returnedUser: any = {
+                id: userResult.id,
+                username: userResult.username,
+                email: userResult.email,
+                role: userResult.role
+            }
+
             res.status(200).send({
                 message: 'User found',
-                user: userResult
+                user: returnedUser
             });
         } else {
             res.status(404).send({ message: 'User not found' });
@@ -92,13 +155,13 @@ const getUserById = async (req: Request, res: Response) => {
 
 const addNewUser = async (req: Request, res: Response) => {
     try {
-        const { username, email, password, role } = req.body;
-        const exitingUsername: User | null = await user.getUserByUsernameOrEmail(username);
+        const { username, email, password } = req.body;
+        // const exitingUsername: User | null = await user.getUserByUsernameOrEmail(username);
         const exitingEmail: User | null = await user.getUserByUsernameOrEmail(email);
 
-        if(exitingUsername) {
-            return res.status(400).send({ message: 'Username already used' });
-        }
+        // if(exitingUsername) {
+        //     return res.status(400).send({ message: 'Username already used' });
+        // }
 
         if(exitingEmail) {
             return res.status(400).send({ message: 'Email already used' });
@@ -109,7 +172,8 @@ const addNewUser = async (req: Request, res: Response) => {
             username: '',
             email: '',
             password: '',
-            role: 'USER'
+            role: 'USER',
+            image_url: ''
         };
 
         if(username) {
@@ -148,19 +212,36 @@ const addNewUser = async (req: Request, res: Response) => {
             return res.status(400).send({ message: 'Password is required' });
         }
 
-        if(role) {
-            if(role === 'USER' || role === 'ADMIN') {
-                newUser.role = role;
-            } else {
-                return res.status(400).send({ message: 'Invalid role' });
-            }
-        }
-
         try {
             const insertedId = await user.addUser(newUser);
+
+            if (!insertedId) {
+                return res.status(500).send({ message: 'Failed to add user' });
+            }
+    
+            newUser.id = insertedId;
+            
+            const secretKey = process.env.JWT_SECRET || 'default_secret_key';
+    
+            const tokenPayload = {
+                id: insertedId,
+                username: newUser.username,
+                email: newUser.email,
+                role: newUser.role
+            };
+    
+            const token = jwt.sign(tokenPayload, secretKey, { expiresIn: '2h' });
+            const connectedUser: any = {
+                id: insertedId,
+                username: newUser.username,
+                role: newUser.role,
+                image_url: newUser.image_url
+            }
+
             res.status(200).send({
                 message: 'User added successfully',
-                userId: insertedId
+                token: token,
+                connectedUser: connectedUser
             });
         } catch (error) {
             console.error('Error adding user:', error);
@@ -175,14 +256,20 @@ const addNewUser = async (req: Request, res: Response) => {
 const updateExistingUser = async (req: Request, res: Response) => {
     try {
         const userId = Number(req.params.id);
-        const { password, role } = req.body;
+
+        if (!((req.isAuth && req.user.id === userId) || (req.isAuth && req.user.role === 'ADMIN'))) {
+            return res.status(400).send({ message: 'You are not authorized to perform this action.' });
+        } 
+
+        const { password, image_url } = req.body;
 
         let updatedUser: User = {
             id: 0,
             username: '',
             email: '',
             password: '',
-            role: ''
+            role: '',
+            image_url: ''
         };
 
         const userResult = await user.getUserById(userId);
@@ -205,12 +292,8 @@ const updateExistingUser = async (req: Request, res: Response) => {
             }
         }
 
-        if(role) {
-            if(role === 'USER' || role === 'ADMIN') {
-                updatedUser.role = role;
-            } else {
-                return res.status(400).send({ message: 'Invalid role' });
-            }
+        if(image_url && image_url.length> 0) {
+            updatedUser.image_url = image_url;
         }
 
         const updated = await user.updateUser(userId, updatedUser);
@@ -232,6 +315,11 @@ const updateExistingUser = async (req: Request, res: Response) => {
 const deleteUserById = async (req: Request, res: Response) => {
     try {
         const userId = Number(req.params.id);
+
+        if (!(req.isAuth && req.user.id === userId)) {
+            return res.status(400).send({ message: 'You are not authorized to perform this action.' });
+        } 
+
         const deleted = await user.deleteUser(userId);
 
         if (deleted) {
@@ -248,4 +336,4 @@ const deleteUserById = async (req: Request, res: Response) => {
     }
 };
 
-export default { login, getAll, getUserById, addNewUser, updateExistingUser, deleteUserById };
+export default { validateToken, login, logout, getAll, getUserById, addNewUser, updateExistingUser, deleteUserById };
